Time: 0 ms (100%), Space: 8 MB (15.97%) - LeetHub

+++ 0795-k-th-symbol-in-grammar/0795-k-th-symbol-in-grammar.cpp
@@ -6,7 +6,7 @@ int solve (int n , int k){
 
         return 0;
     }
-    int mid = (int)(pow(2, n - 1) / 2);
+    int mid = pow(2, n - 1) / 2;
     if(k<=mid){
         return solve(n-1,k);
